ref: move hotkeyInput to utils
